requirements.txt
beautifulsoup4==4.12.2
dominate==2.9.0
flask==2.3.3
genanki==0.13.1
greek-accentuation==1.2.0
lxml==4.9.4
more-itertools==10.1.0
ordered-enum==0.0.8
pyconll==3.2.0
pyCTS==0.2.2
python-box==7.1.1
python-slugify==8.0.1
requests==2.31.0
rich==13.7.0
spacy==3.7.2
tornado==6.2
werkzeug==3.0.1


# dev
djlint==1.34.1
lxml-stubs==0.4.0
mypy==1.7.1
pytest==7.4.3
types-python-slugify==8.0.0.3
watchdog==3.0.0
watchdog[watchmedo]==3.0.0
